Moved mapping functions out to own file so that I can reuse them more.
