+++ src/app/loadout-builder/hooks/useProcess.ts
@@ -10,24 +10,17 @@ import {
   MinMaxIgnored,
   MinMax,
   LockedModBase,
-  statHashToType,
   ModPickerCategories,
 } from '../types';
-import { DimItem, DimSocket, DimSockets, D2Item } from 'app/inventory/item-types';
+import { DimItem } from 'app/inventory/item-types';
+import { ProcessItemsByBucket } from '../processWorker/types';
 import {
-  ProcessItemsByBucket,
-  ProcessItem,
-  ProcessArmorSet,
-  ProcessSocket,
-  ProcessSockets,
-  LockedArmor2ProcessMods,
-  ProcessMod,
-} from '../processWorker/types';
-import {
-  getSpecialtySocketMetadata,
-  getSpecialtySocketMetadataByPlugCategoryHash,
-} from 'app/utils/item-utils';
-import { DestinyItemInvestmentStatDefinition } from 'bungie-api-ts/destiny2';
+  mapDimItemToProcessItem,
+  mapSeasonalModsToProcessMods,
+  getTotalModStatChanges,
+  mapArmor2ModsToProcessMods,
+  hydrateArmorSet,
+} from '../processWorker/mappers';
 
 interface ProcessState {
   processing: boolean;
@@ -39,8 +32,6 @@ interface ProcessState {
   } | null;
   currentCleanup: (() => void) | null;
 }
-
-type ItemsById = { [id: string]: DimItem };
 
 /**
  * Hook to process all the stat groups for LO in a web worker.
@@ -82,7 +73,7 @@ export function useProcess(
     setState({ processing: true, result, currentCleanup: cleanup });
 
     const processItems: ProcessItemsByBucket = {};
-    const itemsById: ItemsById = {};
+    const itemsById: { [id: string]: DimItem } = {};
 
     for (const [key, items] of Object.entries(filteredItems)) {
       processItems[key] = [];
@@ -192,171 +183,4 @@ function createWorker() {
   };
 
   return { worker, cleanup };
-}
-
-// TODO Move all the stuff below here to its own file so the hook specific stuff is clear
-
-function mapDimSocketToProcessSocket(dimSocket: DimSocket): ProcessSocket {
-  return {
-    plug: dimSocket.plug && {
-      stats: dimSocket.plug.stats,
-      plugItemHash: dimSocket.plug.plugItem.hash,
-    },
-    plugOptions: dimSocket.plugOptions.map((dimPlug) => ({
-      stats: dimPlug.stats,
-      plugItemHash: dimPlug.plugItem.hash,
-    })),
-  };
-}
-
-/**
- * Maps the seasonal mods from the PerkPicker to ProcessMods.
- *
- * TODO When ModPicker is fully merged delete this.
- */
-function mapSeasonalModsToProcessMods(lockedSeasonalMods: readonly LockedModBase[]): ProcessMod[] {
-  const metadatas = lockedSeasonalMods.map((mod) => ({
-    mod,
-    metadata: getSpecialtySocketMetadataByPlugCategoryHash(mod.mod.plug.plugCategoryHash),
-  }));
-
-  const modMetadata: ProcessMod[] = [];
-  for (const entry of metadatas) {
-    modMetadata.push({
-      season: entry.metadata?.season,
-      tag: entry.metadata?.tag,
-      energyType: entry.mod.mod.plug.energyCost.energyType,
-      investmentStats: entry.mod.mod.investmentStats.map(({ statTypeHash, value }) => ({
-        statTypeHash,
-        value,
-      })),
-    });
-  }
-
-  return modMetadata;
-}
-
-/**
- * Maps armour 2.0 mods from the ModPicker to ProcessMods.
- */
-function mapArmor2ModsToProcessMods(lockedMods: LockedArmor2ModMap): LockedArmor2ProcessMods {
-  const seasonalMetas = lockedMods.seasonal.map((mod) =>
-    getSpecialtySocketMetadataByPlugCategoryHash(mod.mod.plug.plugCategoryHash)
-  );
-
-  return _.mapValues(lockedMods, (mods) =>
-    mods.map((mod, index) => {
-      const processMod = {
-        energyType: mod.mod.plug.energyCost.energyType,
-        investmentStats: mod.mod.investmentStats,
-      };
-
-      if (mod.category === 'seasonal') {
-        return {
-          ...processMod,
-          season: seasonalMetas[index]?.season,
-          tag: seasonalMetas[index]?.tag,
-        };
-      }
-
-      return processMod;
-    })
-  );
-}
-
-/**
- * This sums up the total stat contributions across mods passed in. These are then applied
- * to the loadouts after all the items base values have been summed. This mimics how seasonal mods
- * effect stat values in game and allows us to do some preprocessing.
- *
- * For the Mod Picker this can be used for seasonal and general mods. For mods in perk picker this is
- * just for the seasonal mods.
- */
-function getTotalModStatChanges(
-  lockedSeasonalMods: readonly { mod: { investmentStats: DestinyItemInvestmentStatDefinition[] } }[]
-) {
-  const totals: { [stat in StatTypes]: number } = {
-    Mobility: 0,
-    Recovery: 0,
-    Resilience: 0,
-    Intellect: 0,
-    Discipline: 0,
-    Strength: 0,
-  };
-
-  for (const mod of lockedSeasonalMods) {
-    for (const stat of mod.mod.investmentStats) {
-      const statType = statHashToType[stat.statTypeHash];
-      if (statType) {
-        totals[statType] += stat.value;
-      }
-    }
-  }
-
-  return totals;
-}
-
-function mapDimSocketsToProcessSockets(dimSockets: DimSockets): ProcessSockets {
-  return {
-    sockets: dimSockets.sockets.map(mapDimSocketToProcessSocket),
-    categories: dimSockets.categories.map((category) => ({
-      categoryStyle: category.category.categoryStyle,
-      sockets: category.sockets.map(mapDimSocketToProcessSocket),
-    })),
-  };
-}
-
-function mapDimItemToProcessItem(dimItem: D2Item): ProcessItem {
-  const { bucket, id, type, name, equippingLabel, basePower, stats } = dimItem;
-
-  const statMap: { [statHash: number]: number } = {};
-  const baseStatMap: { [statHash: number]: number } = {};
-
-  if (stats) {
-    for (const { statHash, value, base } of stats) {
-      statMap[statHash] = value;
-      baseStatMap[statHash] = base;
-    }
-  }
-
-  const modMetadata = getSpecialtySocketMetadata(dimItem);
-
-  return {
-    bucketHash: bucket.hash,
-    id,
-    type,
-    name,
-    equippingLabel,
-    basePower,
-    stats: statMap,
-    baseStats: baseStatMap,
-    sockets: dimItem.sockets && mapDimSocketsToProcessSockets(dimItem.sockets),
-    energyType: dimItem.energy?.energyType,
-    season: modMetadata?.season,
-    compatibleModSeasons: modMetadata?.compatibleTags,
-  };
-}
-
-export function hydrateArmorSet(processed: ProcessArmorSet, itemsById: ItemsById): ArmorSet {
-  const sets: ArmorSet['sets'] = [];
-
-  for (const processSet of processed.sets) {
-    const armor: DimItem[][] = [];
-
-    for (const itemIds of processSet.armor) {
-      armor.push(itemIds.map((id) => itemsById[id]));
-    }
-
-    sets.push({ armor, statChoices: processSet.statChoices });
-  }
-
-  const firstValidSet: DimItem[] = processed.firstValidSet.map((id) => itemsById[id]);
-
-  return {
-    sets,
-    firstValidSet,
-    firstValidSetStatChoices: processed.firstValidSetStatChoices,
-    stats: processed.stats,
-    maxPower: processed.maxPower,
-  };
 }
